requirements.txt
flask==2.0.1
Flask-PyMongo==2.3.0
python-dotenv==0.19.1
flask-jwt-extended==4.3.1
flask-limiter==2.0.4
Werkzeug==2.0.1
bson==0.5.10
flask-cors==3.0.10
dnspython==2.7.0
certifi==2023.7.22
